add lot of changes

+++ web/src/actions/user/user.ts
@@ -1,9 +1,10 @@
-import { ThunkDispatch, ThunkAction } from 'redux-thunk';
-import { User, MutationSignInArgs, MutationSignUpArgs, getApi } from '../../helpers';
-import { RootState } from '../../reducers';
-import { localStorageKeys, routerUri } from '../../config';
 import { push, RouterAction } from 'connected-react-router';
-import { SignInAction } from './signIn'
+import { ThunkDispatch } from 'redux-thunk';
+import { localStorageKeys, routerUri } from '../../config';
+import { getApi, MutationSignUpArgs, User } from '../../helpers';
+import { RootState } from '../../reducers';
+import { enqueueSnackbar } from '../notifier';
+import { SharedActions } from '../shared';
 
 export interface GetCurrentUserActionFetching {
     type: 'GET_CURRENT_USER_REQUEST';
@@ -27,25 +28,24 @@ export interface CurrentUserActionSetAesPassphrase {
 }
 
 export type GetCurrentUserAction =
-    | RouterAction
+    | SharedActions
     | GetCurrentUserActionFetching
     | GetCurrentUserActionSuccess
     | GetCurrentUserActionFailure;
-
 
 export const fetchCurrentUser = (options?: { forceReload: boolean }) => async (
     dispatch: ThunkDispatch<{}, {}, GetCurrentUserAction>,
     getState: () => RootState,
 ) => {
-
     const state = getState();
-    const token = state.currentUser.token
+    const token = state.currentUser.token;
 
     if (!token) {
-        dispatch(push(routerUri.signIn))
+        console.debug('Undefined token. Redirecting to /sign-in')
+        dispatch(push(routerUri.signIn));
         return;
     }
-    const api = getApi({ token })
+    const api = getApi({ token });
     if (!options?.forceReload) {
         if (
             state.currentUser.isFetching ||
@@ -84,18 +84,30 @@ export const setAesPassphrase = (
     aesPassphrase: string,
     shouldSaveToLocalstorage: boolean,
 ) => async (
-    dispatch: ThunkDispatch<{}, {}, CurrentUserActionSetAesPassphrase>,
+    dispatch: ThunkDispatch<{}, {}, CurrentUserActionSetAesPassphrase | SharedActions>,
     getState: () => RootState,
     ) => {
+        const state = getState()
+        const isDifferent = state.currentUser.aesPassphrase !== aesPassphrase
+        if (isDifferent) {
+            if (!window.confirm('Are you sure you want to change your passphrase?')) {
+                return
+            }
+        }
         dispatch({ type: 'SET_AES_PASSPHRASE', user: { aesPassphrase } });
         if (shouldSaveToLocalstorage) {
             localStorage.setItem(localStorageKeys.aesPassphrase, aesPassphrase);
         }
+        dispatch(enqueueSnackbar({
+            message: 'AES Passphrase updated',
+            options: { variant: 'success' }
+        }))
+        if (isDifferent) {
+            setTimeout(() => window.location.reload(), 0)
+        }
     };
 
-
 export const signUp = async (variables: MutationSignUpArgs['input']) => {
-
     // const { signUp: token } = await api.signUp({ input: variables });
     // if (token) {
     //     localStorage.setItem(localStorageKeys.token, token);
