+++ web/src/containers/NoteListContainer.tsx
@@ -1,64 +1,22 @@
 import { Button, CircularProgress } from '@material-ui/core';
 import React, { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { deleteNote, fetchCurrentUserNotes, updateNote } from '../../actions';
-import { NoteList, CreateNoteFormValues } from '../../components';
-import { RootState } from '../../reducers';
+import { deleteNote, updateNote } from '../actions';
+import { NoteList, CreateNoteFormValues } from '../components';
+import { RootState } from '../reducers';
 
-interface NoteListContainerProps {}
-
-export const getTagsIdFromSearch = (
-    decryptedTags: any[],
-    searchFilter?: string | null,
-): any[] => {
-    if (!searchFilter) {
-        return [];
-    }
-    const searchFilterTags = searchFilter
-        ? decodeURIComponent(searchFilter || '')
-              .split(',')
-              .map(word => word.trim())
-        : [];
-    const tagsIdFilter = decryptedTags
-        ? decryptedTags
-              .filter((tag: any) => searchFilterTags.includes(tag.label))
-              .map((tag: any) => tag.id)
-        : [];
-
-    return tagsIdFilter;
-};
+interface NoteListContainerProps {
+    loadMore(): void;
+    displayedNotes: RootState['currentUserNotes'] | RootState['searchNotes'];
+}
 
 export const NoteListContainer: React.SFC<NoteListContainerProps> = props => {
     const [editingNoteId, setEditingNoteId] = useState('');
     const dispatch = useDispatch();
 
-    const currentUserNotes = useSelector(
-        (state: RootState) => state.currentUserNotes,
-    );
     const currentUserTags = useSelector(
         (state: RootState) => state.currentUserTags,
     );
-
-    const searchNotes = useSelector((state: RootState) => state.searchNotes);
-
-    const displayedNotes = searchNotes.searchValue
-        ? searchNotes
-        : currentUserNotes;
-
-    const loadMore = () => {
-        if (!currentUserNotes.isFetching && currentUserNotes.hasMore) {
-            dispatch(
-                fetchCurrentUserNotes({
-                    forceReload: true,
-                    variables: {
-                        limit: 100,
-                        skip: currentUserNotes.notes.length,
-                    },
-                }),
-            );
-        }
-    };
-
     return (
         <div>
             <div>
@@ -78,9 +36,11 @@ export const NoteListContainer: React.SFC<NoteListContainerProps> = props => {
                         tags={currentUserTags.tags}
                         isTagLoading={currentUserTags.isFetching}
                         onDeleteClick={noteId => {
-                            dispatch(deleteNote(noteId));
+                            if (window.confirm('Delete note?')) {
+                                dispatch(deleteNote(noteId));
+                            }
                         }}
-                        notes={displayedNotes.notes.map(note => {
+                        notes={props.displayedNotes.notes.map(note => {
                             const tags = note.tags
                                 .map(
                                     tag =>
@@ -98,13 +58,12 @@ export const NoteListContainer: React.SFC<NoteListContainerProps> = props => {
                     />
                 }
             </div>
-            {(currentUserNotes.isFetching || currentUserTags.isFetching) && (
-                <CircularProgress />
-            )}
-            {!currentUserNotes.isFetching && currentUserNotes.hasMore && (
+            {(props.displayedNotes.isFetching ||
+                currentUserTags.isFetching) && <CircularProgress />}
+            {!props.displayedNotes.isFetching && props.displayedNotes.hasMore && (
                 <div>
                     <Button
-                        onClick={loadMore}
+                        onClick={props.loadMore}
                         style={{ padding: 30 }}
                         fullWidth
                     >
